+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ### v0.next
 
+### v0.7.3
+
+* Fix: Nested `@rest(…)` calls with nested `@export(as:…)` directives should keep their contexts distinct in order to work. [#204](https://github.com/apollographql/apollo-link-rest/pull/204)
+
+### v0.7.2
+
+* Fix: FileList/File aren't available in react-native causing crashes. [#200](https://github.com/apollographql/apollo-link-rest/pull/200)
+
+### v0.7.1
+
+* Fix: Duplicated Content Type Header [#188](https://github.com/apollographql/apollo-link-rest/pull/188)
+* Fix: FileList Support [#183](https://github.com/apollographql/apollo-link-rest/pull/183)
+* Fix: Default Empty object when creating headers [#178](https://github.com/apollographql/apollo-link-rest/pull/178)
+* Body-containing Queries [#173](https://github.com/apollographql/apollo-link-rest/pull/173)
+
 ### v0.7.0 - Breaking!
 
 #### Breaking changes around `responseTransformer!`